--- Epsilon.Abstractions/Component/KpiMatrix.cs
@@ -1,7 +1,0 @@
-﻿namespace Epsilon.Abstractions.Component;
-
-public record KpiMatrix(
-    IEnumerable<KpiMatrixOutcome> Outcomes,
-    IEnumerable<KpiMatrixAssignment> Assignments
-);
-
