--- Epsilon.Abstractions/Component/KpiMatrixAssignment.cs
@@ -1,6 +1,0 @@
-namespace Epsilon.Abstractions.Component;
-
-public record KpiMatrixAssignment(
-    string Name,
-    IEnumerable<KpiMatrixOutcome> Outcomes
-);
