+++ Epsilon.Abstractions/Component/KpiMatrixComponent/KpiMatrixModule.cs
@@ -1,4 +1,4 @@
-namespace Epsilon.Abstractions.Component;
+namespace Epsilon.Abstractions.Component.KpiMatrixComponent;
 
 public record KpiMatrixModule(
     string Name,
